Disable dired commands when on Vim search/command
This is the equivalent of `!findWidgetVisible` but for users of
the Vim extension.

This still works if the vim extension is not installed.

Granted, some of these commands are not ideal for Vim (eg `u`, `g`,
`q`) but I still find that I get value from using this extension via the
Vim mode.